# api/openapi.v1.yaml
...
components:
  schemas:
...
    CheckIn:
      additionalProperties: false
      description: 'Raw current-state submission.


        ``extra="forbid"`` keeps the input surface bounded: an unknown field is a

        422 rather than a silently ignored value that changes nothing.'
      properties:
        available_minutes:
          enum:
          - 3
          - 5
          - 10
          - 15
          - 20
          title: Available Minutes
          type: integer
        energy:
          maximum: 10.0
          minimum: 0.0
          title: Energy
          type: integer
        experience_level:
          $ref: '#/components/schemas/ExperienceLevel'
        goal:
          $ref: '#/components/schemas/Goal'
        mental_activity:
          maximum: 10.0
          minimum: 0.0
          title: Mental Activity
          type: integer
        sleepiness:
          maximum: 10.0
          minimum: 0.0
          title: Sleepiness
          type: integer
        stress:
          maximum: 10.0
          minimum: 0.0
          title: Stress
          type: integer
      required:
      - goal
      - stress
      - energy
      - mental_activity
      - sleepiness
      - available_minutes
      - experience_level
      title: CheckIn
      type: object
    CheckInResponse:
      additionalProperties: false
      properties:
        check_in:
          $ref: '#/components/schemas/CheckIn'
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          format: uuid
          title: Id
          type: string
      required:
      - id
      - created_at
      - check_in
      title: CheckInResponse
      type: object
...
    ErrorResponse:
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
      - error
      title: ErrorResponse
      type: object
...
    ExposureRequest:
      additionalProperties: false
      description: Reports that a variant was actually shown to the user.
      properties:
        context:
          maxLength: 64
          minLength: 1
          title: Context
          type: string
        experiment_id:
          maxLength: 64
          title: Experiment Id
          type: string
      required:
      - experiment_id
      - context
      title: ExposureRequest
      type: object
    ExposureResponse:
      additionalProperties: false
      properties:
        context:
          title: Context
          type: string
        experiment_id:
          title: Experiment Id
          type: string
        recorded:
          title: Recorded
          type: boolean
        variant:
          title: Variant
          type: string
      required:
      - experiment_id
      - variant
      - context
      - recorded
      title: ExposureResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
info:
  description: 'Adaptive Meditation Program001 vertical slice: check-in, deterministic practice recommendation,
    session protocol and outcome feedback. Wellness product; it does not diagnose or treat any condition.'
  title: Adaptive Meditation API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/check-ins:
    post:
      description: 'Guest-first: no account is required and no user identity is created.


        A guest id, if sent, is recorded so the caller can find and delete this data

        later. Without one the check-in is still accepted; it simply has no owner.'
      operationId: createCheckIn
      parameters:
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: Record a current-state check-in
      tags:
      - v1
...
  /v1/experiments/exposures:
    post:
      description: 'Exposure, as distinct from assignment.


        Idempotent per context, so re-opening the same screen does not double-count.

        ``recorded`` says whether this call created the exposure or found one.'
      operationId: recordExperimentExposure
      parameters:
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExposureRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExposureResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Guest identity required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: Record that an experiment variant was actually shown
      tags:
      - v1
...
  /v1/recommendations:
    post:
      description: 'The recommendation itself is pure computation.


        A session is opened only to record the presentation-experiment assignment,

        and only when the caller sent a guest identity. Without one this endpoint

        still touches no database at all, which is what keeps it answerable with the

        database down.'
      operationId: createRecommendation
      parameters:
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: Deterministic practice recommendation
      tags:
      - v1
  /v1/recommendations/candidates:
    post:
      description: 'Not part of the normal client flow.


        The ordinary recommendation endpoint returns the winner only; this exposes

        the ranking behind it for offline evaluation. Scores are ordinal ranking

        aids, not probabilities.'
      operationId: createRecommendationCandidates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateListResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: Full scored candidate list (debug and offline evaluation)
      tags:
      - v1
  /v1/sessions:
    post:
      operationId: createSession
      parameters:
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflicting state
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: Create a session from a stored check-in
      tags:
      - v1
...
  /v1/sessions/{session_id}:
    get:
      operationId: getSession
      parameters:
      - in: path
        name: session_id
        required: true
        schema:
          format: uuid
          title: Session Id
          type: string
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Guest identity required
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Read a session, including its recovery point
      tags:
      - v1
  /v1/sessions/{session_id}/events:
    post:
      description: 'Batched so a session does not make a round trip per segment.


        Idempotent per (session, sequence): a client that retries after losing its

        connection re-sends the batch and gets the same journal, not a doubled one.'
      operationId: appendSessionEvents
      parameters:
      - in: path
        name: session_id
        required: true
        schema:
          format: uuid
          title: Session Id
          type: string
      - in: header
        name: X-Guest-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Guest-Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionEventBatchRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEventBatchResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Guest identity required
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: Append playback events
      tags:
      - v1
  /v1/sessions/{session_id}/feedback:
    post:
      operationId: submitSessionFeedback
      parameters:
      - in: path
        name: session_id
        required: true
        schema:
          format: uuid
          title: Session Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionFeedbackRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation error
      summary: Attach before/after feedback to a session
      tags:
      - v1
...
  /v1/sessions/{session_id}/start:
    post:
      operationId: startSession
      parameters:
      - in: path
        name: session_id
        required: true
        schema:
          format: uuid
          title: Session Id
          type: string
      responses:
        '204':
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Mark a session as started
      tags:
      - v1